+++ packages/opentelemetry-sdk-trace-web/src/utils.ts
@@ -15,10 +15,6 @@ import {
   timeInputToHrTime,
   urlMatches,
 } from '@opentelemetry/core';
-import {
-  ATTR_HTTP_RESPONSE_CONTENT_LENGTH,
-  ATTR_HTTP_RESPONSE_CONTENT_LENGTH_UNCOMPRESSED,
-} from './semconv';
 
 // Used to normalize relative URLs
 let urlNormalizingAnchor: HTMLAnchorElement | undefined;
@@ -67,14 +63,13 @@ export function addSpanNetworkEvent(
 }
 
 /**
- * Helper function for adding network events and content length attributes.
+ * Helper function for adding network events to a span.
  */
 export function addSpanNetworkEvents(
   span: api.Span,
   resource: PerformanceEntries,
   ignoreNetworkEvents = false,
-  ignoreZeros?: boolean,
-  skipOldSemconvContentLengthAttrs?: boolean
+  ignoreZeros?: boolean
 ): void {
   if (ignoreZeros === undefined) {
     ignoreZeros = resource[PTN.START_TIME] !== 0;
@@ -95,24 +90,6 @@ export function addSpanNetworkEvents(
     addSpanNetworkEvent(span, PTN.REQUEST_START, resource, ignoreZeros);
     addSpanNetworkEvent(span, PTN.RESPONSE_START, resource, ignoreZeros);
     addSpanNetworkEvent(span, PTN.RESPONSE_END, resource, ignoreZeros);
-  }
-
-  if (!skipOldSemconvContentLengthAttrs) {
-    // This block adds content-length-related span attributes using the
-    // *old* HTTP semconv (v1.7.0).
-    const encodedLength = resource[PTN.ENCODED_BODY_SIZE];
-    if (encodedLength !== undefined) {
-      span.setAttribute(ATTR_HTTP_RESPONSE_CONTENT_LENGTH, encodedLength);
-    }
-
-    const decodedLength = resource[PTN.DECODED_BODY_SIZE];
-    // Spec: Not set if transport encoding not used (in which case encoded and decoded sizes match)
-    if (decodedLength !== undefined && encodedLength !== decodedLength) {
-      span.setAttribute(
-        ATTR_HTTP_RESPONSE_CONTENT_LENGTH_UNCOMPRESSED,
-        decodedLength
-      );
-    }
   }
 }